LVPN-3630: Move file watcher out of internal pkg (#1454)
