+++ internal/filesystem.go
@@ -15,7 +15,6 @@ import (
 	"strings"
 
 	"github.com/NordSecurity/nordvpn-linux/log"
-	"github.com/fsnotify/fsnotify"
 	"golang.org/x/sys/unix"
 )
 
@@ -617,28 +616,6 @@ func isProcessRunning(executablePath string, readdir readdirFunc, readfile readf
 	}
 
 	return false, nil
-}
-
-// GetFileWatcher returns a fsnotify file watcher that is monitoring files provided in pathsToMonitor
-func GetFileWatcher(pathsToMonitor ...string) (watcher *fsnotify.Watcher, err error) {
-	watcher, err = fsnotify.NewWatcher()
-	if err != nil {
-		return nil, fmt.Errorf("creating new watcher: %w", err)
-	}
-
-	defer func() {
-		if err != nil && watcher != nil {
-			_ = watcher.Close()
-		}
-	}()
-
-	for _, file := range pathsToMonitor {
-		if err := watcher.Add(file); err != nil {
-			return nil, fmt.Errorf("adding group file to watcher: %w", err)
-		}
-	}
-
-	return watcher, nil
 }
 
 // UserLogOutput opens logFileName in the user's cache directory and returns it.
